+++ src/main/java/dao/TrackerDao.java
@@ -22,7 +22,7 @@ public class TrackerDao extends DAO<Tracker>{
     }
 
     @Override
-    public Tracker findById(Long id) {
+    public Tracker findById(Integer id) {
         return super.findById(id); //To change body of generated methods, choose Tools | Templates.
     }
