+++ src/main/java/service/TrackerService.java
@@ -23,7 +23,7 @@ public class TrackerService extends Service<Tracker>{
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Integer id) {
         super.delete(id); //To change body of generated methods, choose Tools | Templates.
     }
 
@@ -33,7 +33,7 @@ public class TrackerService extends Service<Tracker>{
     }
 
     @Override
-    public Tracker findById(Long id) {
+    public Tracker findById(Integer id) {
         return super.findById(id); //To change body of generated methods, choose Tools | Templates.
     }
